refactor tsv modal options
